Made get_network_stats() a utility
